backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
firebase-admin==6.2.0
pydantic==2.5.1
pydantic[email]==2.5.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
alembic==1.12.1
slowapi==0.1.8
python-multipart==0.0.6
requests==2.31.0
httpx==0.25.1
pytest==7.4.3
pytest-asyncio==0.21.1
coverage==7.3.2
black==23.10.1
isort==5.12.0
flake8==6.1.0
mypy==1.7.0
